LTSINSCRIP-41 attr_accessible set on all models that have DBs backing them
This closes this issue, but should NOT be considered a end-all fix
for mass assignment vulnerabilities across Inscriptio.  All relevant models
have a whitelist set up, but these whitelists may be overpermissive - they're
basically developed based on what will allow current functioning.

Note that the Report and Search models are both empty and not backed by the DB.
They aren't going to be DB-backed (it wouldn't make much sense, unless we were
using the DB as a cache layer for query results), but there's a lot of logic in
their views and controllers that should be moved up from the controllers and
views.