LTSINSCRIP-21 Clear functionality implemented via Paranoid, old status-based stuff cleared out.
Archived stuff is set as expired but not cleared (deleted) - this migration is kind of terrible
and, like its predecessor, covers functionality that I'm really not convinced belongs in a migration,
but for which there is no better place.

Currently, admins can kill cleared things permanently by visiting reservations/ID of the cleared
reservation and deleting it.  This isn't intended for use, but is handled sanely in case there's a
future need.  Reviving cleared reservations ALSO doesn't really exist.  Everything else that is paranoid
has no UI either way - they're just silently soft-deleted.