LTSINSCRIP-76 , LTSINSCRIP-75 (sort of) and more RailsUJS stuff
This branch consists of a marginally cleaner and more performant rewrite of the
functionality on the "Reports" admin page.  I've located it at Reservation#index, because
there's no user listing of reservations, and it seemed to fit.

Additionally, delete and update functionality for reservations is handled
via ajax (Rails UJS, to be specific).  I'm doing it in the low-pass way, by letting
the Rails UJS helper eval returned JS.  We have CSRF protection enabled, so it's not
insecure, but arguably it could be done more cleanly by hanging off the 'ajax:success'
and 'ajax:failure' methods.  Not gonna, for now.